+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 Unreleased
 ==========
+* feat: Add AliasContent admin changelist Category filter
 
 1.5.0 (2022-04-26)
 ==================